+++ xenon-common/src/test/java/com/vmware/xenon/services/common/authn/TestBasicAuthenticationService.java
@@ -118,8 +118,7 @@ public class TestBasicAuthenticationService {
                                         return;
                                     }
                                     String authHeader = o
-                                            .getResponseHeaders()
-                                            .get(
+                                            .getResponseHeader(
                                                     BasicAuthenticationService.WWW_AUTHENTICATE_HEADER_NAME);
                                     if (authHeader == null
                                             || !authHeader